+++ libs/core/custom-elements-manifest.config.mjs
@@ -45,26 +45,6 @@ export default {
         })
       },
     },
-    {
-      name: 'green-react-event-names',
-      analyzePhase({ ts, node, moduleDoc }) {
-        switch (node.kind) {
-          case ts.SyntaxKind.ClassDeclaration: {
-            const className = node.name.getText()
-            const classDoc = moduleDoc?.declarations?.find(
-              (declaration) => declaration.name === className,
-            )
-
-            if (classDoc?.events) {
-              classDoc.events.forEach((event) => {
-                event.reactName = `on${pascalCase(event.name)}`
-                event.eventName = `${pascalCase(event.name)}Event`
-              })
-            }
-          }
-        }
-      },
-    },
   ],
 }