+++ libs/core/src/components/button/button.component.ts
@@ -190,13 +190,11 @@ class Button extends GdsFormControlElement<any> {
   }
 
   #handleClick = (e: MouseEvent) => {
-    this.dispatchEvent(
-      new CustomEvent('gds-click', {
-        bubbles: true,
-        composed: true,
-        detail: e,
-      }),
-    )
+    this.dispatchCustomEvent('gds-click', {
+      bubbles: true,
+      composed: true,
+      detail: e,
+    })
 
     if (this.form && !this.#isLink) {
       if (this.type === 'submit') {