+++ libs/core/src/components/checkbox/checkbox-group/checkbox-group.component.ts
@@ -176,12 +176,10 @@ class CheckboxGroup extends GdsFormControlElement<string[]> {
 
   #dispatchInputEvent() {
     this.updateComplete.then(() =>
-      this.dispatchEvent(
-        new Event('input', {
-          bubbles: true,
-          composed: true,
-        }),
-      ),
+      this.dispatchStandardEvent('input', {
+        bubbles: true,
+        composed: true,
+      }),
     )
   }
 
@@ -189,8 +187,7 @@ class CheckboxGroup extends GdsFormControlElement<string[]> {
     return html`<gds-form-control-footer
       id="footer"
       class="size-${this.size}"
-      .validationMessage=${this.invalid &&
-      (this.errorMessage || this.validationMessage)}
+      .errorMessage=${this.invalid ? this.errorMessage : undefined}
     >
     </gds-form-control-footer>`
   }