+++ libs/core/src/components/radio/radio-group/radio-group.component.ts
@@ -144,20 +144,16 @@ class RadioGroup extends GdsFormControlElement<string> {
 
   #dispatchChangeEvents() {
     this.updateComplete.then(() =>
-      this.dispatchEvent(
-        new Event('change', {
-          composed: true,
-          bubbles: true,
-        }),
-      ),
+      this.dispatchStandardEvent('change', {
+        composed: true,
+        bubbles: true,
+      }),
     )
     this.updateComplete.then(() =>
-      this.dispatchEvent(
-        new Event('input', {
-          bubbles: true,
-          composed: true,
-        }),
-      ),
+      this.dispatchStandardEvent('input', {
+        bubbles: true,
+        composed: true,
+      }),
     )
   }
 
@@ -280,8 +276,7 @@ class RadioGroup extends GdsFormControlElement<string> {
     return html` <gds-form-control-footer
       id="footer"
       class="size-${this.size}"
-      .validationMessage=${this.invalid &&
-      (this.errorMessage || this.validationMessage)}
+      .errorMessage=${this.invalid ? this.errorMessage : undefined}
     >
     </gds-form-control-footer>`
   }