+++ libs/tokens/style-dictionary.build.ts
@@ -62,7 +62,7 @@ await Promise.all(
               {
                 destination: `variables.base.css`,
                 format: 'css/only-variables',
-                filter: 'no-colors-no-ref',
+                filter: 'base-tokens',
                 options: {
                   usesDtcg: true,
                 },
@@ -84,6 +84,11 @@ await Promise.all(
                   colorScheme: colorScheme,
                 },
               },
+              {
+                destination: `variables.shadows.css`,
+                format: 'css/shadows',
+                filter: 'is-shadow',
+              },
             ],
           },
           css: {
@@ -104,7 +109,7 @@ await Promise.all(
               {
                 destination: `variables.base.css`,
                 format: 'css/variables',
-                filter: 'no-colors-no-ref',
+                filter: 'base-tokens',
               },
               {
                 destination: `colors.ref.css`,
@@ -137,7 +142,7 @@ await Promise.all(
               {
                 destination: `variables.base.scss`,
                 format: 'scss/variables',
-                filter: 'no-colors-no-ref',
+                filter: 'base-tokens',
               },
               {
                 destination: `_mixin.colors.ref.scss`,